Move heat equation in separate include directory
